requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==4.0.4
django-debug-toolbar==3.4.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
idna==3.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
packaging==21.3
Pillow==9.1.1
prompt-toolkit==3.0.30
psycopg2-binary==2.9.3
PyJWT==2.4.0
pyparsing==3.0.9
python-dotenv==0.20.0
pytz==2022.1
redis==4.3.4
requests==2.28.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.1.2
wrapt==1.14.1
WTForms==3.0.1
zipp==3.8.0
